+++ basicBlock.hh
@@ -121,7 +121,7 @@ public:
     return hasRegion;
   }
   bool hasJR(bool isRet=false) const;
-  bool hasTermDirectBranchOrJump(uint32_t &target, uint32_t &fallthru) const;
+  bool hasTermDirectBranchOrJump(uint64_t &target, uint64_t &fallthru) const;
   bool fallsThru() const;
   bool hasJAL() const {
     return hasjal;
